Stable branch with giant chocolate additions
So something with my fixed width types made a few things not really work so great, so I went and looked at what choco does.  For now, I've gone back to these old types just because it preserves functionality.  Will have to do things slowly to get it working right.

+++ src/engine/net_defs.h
@@ -38,9 +38,9 @@ typedef struct _net_context_s net_context_t;
 struct _net_packet_s
 {
 	byte* data;
-	size_t len;
-	size_t alloced;
-	uint32_t pos;
+	unsigned int len;
+	unsigned int alloced;
+	unsigned int pos;
 };
 
 struct _net_module_s
@@ -154,8 +154,8 @@ typedef struct
 
 typedef struct
 {
-	int32_t latency;
-	uint32_t seq;
+	int latency;
+	unsigned int seq;
 	boolean playeringame[MAXPLAYERS];
 	net_ticdiff_t cmds[MAXPLAYERS];
 } net_full_ticcmd_t;
